Removed apostrophe and updated listing structure
"Its" denotes possession without the apostrophe. "It's" means "it is".

"clean elegant syntax" ought to be "clean, elegant syntax"; however, if we were to do that, then the rest of the commas for the following list would need to be semicolons, so I used a conjunction ("and") instead of the comma.